Initial commit
At this point, there is the JS that adds an address textfield, parses
it into parts, and assigns the parts to other fields according to
a map of part name to field label. All this plugs into the
`woocommerce-custom-fields-txt` plugin. There's also the PHP that
adds 2 fields for simple products - one to enable address parsing,
and one for the map (currently, JSON string). It enqueues the JS,
and gives it the map.